requirements.txt
python-telegram-bot[job-queue]
requests
pandas
plotly
dash
python-dotenv